ingestion/batch/requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
pyarrow>=14.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
pyyaml>=6.0
